Revert "Added upload/download popups to admin view (#12)"
This reverts commit 174333ae2ccefd4c219af87eb2fcf54e9a3823dc.
