--- socialservice/src/app/alumno/components/ui/loco.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const loco = () => {
-  return (
-    <div>loco</div>
-  )
-}
-
-export default loco
