+++ socialservice/src/components/custom/DetalleProyecto.tsx
