+++ socialservice/src/components/custom/PostularseButton.tsx
@@ -11,7 +11,7 @@ export const PostularseButton = ({ texto}: Props) => {
     const router = useRouter();
   
     const handleNavigation = () => {
-      router.push("/alumno/explorar/proyecto/formulario");
+      router.push("/explorar/proyecto/formulario");
     };
   
   return (
